old-tests/submission/Microsoft/async/async_003.htm

<!DOCTYPE html>
<html>
    <head>
        <title>An async script does not block the parser while downloading</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test ensures an async script does not block the parser while downloading." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" /> 
        <link rel="help" href="http://dev.w3.org/html5/spec/scripting-1.html#attr-script-async"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">        
        
  
        var t = async_test("An async script does not block the parser while downloading");
  
        function timeout()
        {
            t.step(function(){ assert_equals(document.getElementById("testresult").innerHTML, "21")});
            t.done();
        }
  
        var timer = setTimeout(timeout, 5000);
        
        // This test uses a script (log.py) to change script order execution
        // The php script takes two parameters 
        //  * The first is for a the delay (sec=1 == sleep(1) in php)
        //  * The second is passed as a parameter to 'log()'
        //
        // The log function just appends to the testresult tag.
        //
        // In this case the log.py script tag has the 'async' attribute so the parser doesn't block
        // Because of the sleep in the php file the 'log(2)' function gets called and adds a '2' to the testresult tag
        // Then the php file actually gets downloaded and runs log(1) which appends a '1' to the testresult tag
        // 
        //  Thus the assert above is to check that the innerHTML == 21
        //
        
        function log(text) 
        {
            var textNode = document.createTextNode(text);
            document.getElementById("testresult").appendChild(textNode);
        }
        
        </script>
        
        <span style="display : none" id="testresult"></span>

        <script src="log.py?sec=3&id=1" async></script>
        <script>
            log('2');
        </script>
    </body>
</html>
